+++ app/dashboard/invoices/[id]/edit/page.tsx
@@ -1,4 +1,4 @@
-import Form from '@/app/ui/invoices/edit-form';
+import Form from '@/src/components/molecules/edit-form';
 import Breadcrumbs from '@/app/ui/invoices/breadcrumbs';
 import { fetchInvoiceById, fetchCustomers } from '@/app/lib/data';
 import { notFound } from 'next/navigation';
